backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
SQLAlchemy>=2.0
asyncpg>=0.29
alembic>=1.13
httpx>=0.27
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
cryptography>=42.0
openai>=1.40
orjson>=3.10
python-multipart>=0.0.9
